refactored some file functions
